// libs/common/src/model/dto/device/move/description/move.description.ts
export enum move_description {
  COMMAND = '이동 명령',
  ID = '요청한 명령의 ID값입니다. request시 자동 생성됩니다.',
  GOALID = 'goal 이동 시, 목표지점의 노드 ID를 입력하세요.',
  METHOD = '주행 방식. 기본값은 pp(point to point)방식입니다.',
  PRESET = '속도 프리셋. 기본값은 0이며 현재 지원하지 않습니다.',
  RESULT = '요청한 명령에 대한 결과입니다. accept, reject, success, fail 등 명령에 대해 다양한 값이 존재합니다.',
  MESSAGE = 'result값이 reject, fail 인 경우 SLAMNAV에서 보내는 메시지 입니다.',
  X = 'target 이동 시, 목표지점의 지도상 x좌표를 입력하세요. 단위는 [m] 입니다.',
  Y = 'target 이동 시, 목표지점의 지도상 y좌표를 입력하세요. 단위는 [m] 입니다.',
  Z = 'target 이동 시, 목표지점의 지도상 z좌표를 입력하세요. 특정 모델만 사용합니다. 기본값 0으로 입력해주세요. 단위는 [m] 입니다.',
  RZ = 'target 이동 시, 목표지점에서 로봇의 z축 방향값을 입력하세요. 단위는 [deg] 입니다.',
  VX = 'jog 이동 시, 로봇의 x방향 속도를 입력하세요. 단위는 [m/s] 입니다.',
  VY = 'jog 이동 시, 로봇의 y방향 속도를 입력하세요. 특정 모델만 사용합니다. 기본값 0으로 입력해주세요. 단위는 [m/s] 입니다.',
  WZ = 'jog 이동 시, 로봇의 z축 회전 속도를 입력하세요. 단위는 [deg/s] 입니다.',
}